First attempt at actual super-readme
But it's very long, and the hierarchy is all messed up.
To fix the hierarchy, I think I need to change the base readmes so that they begin with a level three header.
If the hierarchy doesn't help the length, it's probably best to break this up into a set of pages.

Next step after that is adding top level descriptions to each collection, and removing redundancy.